sqlbase,util/sort: move sort implementation to a util package
Release note: None
