+++ pkg/util/sort/sort.go
@@ -15,7 +15,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package sqlbase
+package sort
 
 import (
 	"sort"
